Support seek() in open files by buffering using SpooledTemporaryFile
Some of the built in Django features expects to be able to `seek()` in file
objects coming from the storage API.

To allow for seeking this patch downloads the whole file in the `open()` function
before sending it along.

If the opened file is large and/or low latency by streaming is required the
application should probably make use of the Minio client directly instead.