Support MINIO_STORAGE_*_URL when generating presigned URLs
This preserves the existing semantics where the a bucket name is expected to be
included within the base_url. It also supports cases where this new bucket name
is different, which is expected by some existing test cases.

This adds a test case to cover the previously-missing case of presigned URLs
without a base_url.

This also improves some existing test assertions to use stricter regexes.

Finally, the changes a test case (and drops support) where an illegal base_url
(with a space character in the endpoint) is used, since Minio will raise an
"InvalidEndpointError: message: Hostname does not meet URL standards."
when initialized with this endpoint. Spaces in key names are still supported
via URL encoding.